Move util layer test into test_fields
